Admin ability to grant access to other users, other minor changes

+++ Controllers/UpdateController.cs
@@ -20,6 +20,12 @@ namespace AssessTracker.Controllers
         [HttpGet]
         [Route("updateAllKids")]
         public IActionResult updateAllKids(){
+            if(HttpContext.Session.GetInt32("Permission") < 5){
+                return RedirectToAction("dashboard", "Dashboard");
+            }
+            if(HttpContext.Session.GetInt32("Permission") == null){
+                return RedirectToAction("Index", "Home");
+            }
             HttpContext.Session.Remove("kidId");
             if(HttpContext.Session.GetString("sort") == "sortLast"){
                 ViewBag.allKids = _context.Kids.Include(kid => kid.Teacher).OrderBy(l => l.LastName);
@@ -78,6 +84,12 @@ namespace AssessTracker.Controllers
         [HttpGet]
         [Route("deactivateKid/{id}")]
         public IActionResult deactivateKid(int id){
+            if(HttpContext.Session.GetInt32("Permission") < 5){
+                return RedirectToAction("dashboard", "Dashboard");
+            }
+            if(HttpContext.Session.GetInt32("Permission") == null){
+                return RedirectToAction("Index", "Home");
+            }
             Kid kid = _context.Kids.SingleOrDefault(x => x.id == id);
             kid.Active = false;
             _context.SaveChanges();
@@ -87,6 +99,12 @@ namespace AssessTracker.Controllers
         [HttpGet]
         [Route("activateKid/{id}")]
         public IActionResult activateKid(int id){
+            if(HttpContext.Session.GetInt32("Permission") < 5){
+                return RedirectToAction("dashboard", "Dashboard");
+            }
+            if(HttpContext.Session.GetInt32("Permission") == null){
+                return RedirectToAction("Index", "Home");
+            }
             Kid kid = _context.Kids.SingleOrDefault(x => x.id == id);
             kid.Active = true;
             _context.SaveChanges();
@@ -96,6 +114,12 @@ namespace AssessTracker.Controllers
         [HttpGet]
         [Route("updateAssessmentPage/{id}")]
         public IActionResult updateAssessmentPage(int id){
+            if(HttpContext.Session.GetInt32("Permission") < 5){
+                return RedirectToAction("dashboard", "Dashboard");
+            }
+            if(HttpContext.Session.GetInt32("Permission") == null){
+                return RedirectToAction("Index", "Home");
+            }
             ViewBag.singleKid = _context.Kids.Where(kid => kid.id == id).Include(t => t.Teacher).ToArray();
             List<DateTaken> DateTakens = _context.DateTaken.Where(x => x.KidId == id).Include(a => a.Assessment).OrderByDescending(d => d.Date).ToList();
             ViewBag.assessmentsTaken = DateTakens;
@@ -117,6 +141,12 @@ namespace AssessTracker.Controllers
         [HttpPost]
         [Route("updateAssessments")]
         public IActionResult updateAssessments(DateTakenViewModel model){
+            if(HttpContext.Session.GetInt32("Permission") < 5){
+                return RedirectToAction("dashboard", "Dashboard");
+            }
+            if(HttpContext.Session.GetInt32("Permission") == null){
+                return RedirectToAction("Index", "Home");
+            }
             if(!ModelState.IsValid){
                 int? id = HttpContext.Session.GetInt32("kidId");
                 ViewBag.singleKid = _context.Kids.Where(kiddo => kiddo.id == id).Include(t => t.Teacher).ToArray();
@@ -154,6 +184,12 @@ namespace AssessTracker.Controllers
         [HttpGet]
         [Route("updateSingleKidPage/{id}")]
         public IActionResult updateSingleKidPage (int id){
+            if(HttpContext.Session.GetInt32("Permission") < 9){
+                return RedirectToAction("updateAllKids");
+            }
+            if(HttpContext.Session.GetInt32("Permission") == null){
+                return RedirectToAction("Index", "Home");
+            }
             List<Kid> kid = _context.Kids.Where(x => x.id == id).Include(k => k.Teacher).ToList();
             ViewBag.kid = kid;
             ViewBag.teachers = _context.Teachers;
@@ -163,6 +199,12 @@ namespace AssessTracker.Controllers
         [HttpPost]
         [Route("updateSingleKid/{id}")]
         public IActionResult updateSingleKid (int id, KidViewModel model){
+            if(HttpContext.Session.GetInt32("Permission") < 9){
+                return RedirectToAction("dashboard", "Dashboard");
+            }
+            if(HttpContext.Session.GetInt32("Permission") == null){
+                return RedirectToAction("Index", "Home");
+            }
             if(!ModelState.IsValid){
                 List<Kid> kid = _context.Kids.Where(x => x.id == id).Include(k => k.Teacher).ToList();
                 ViewBag.kid = kid;
@@ -182,6 +224,12 @@ namespace AssessTracker.Controllers
         [HttpGet]
         [Route("deleteKidPage/{id}")]
         public IActionResult deleteKidPage (int id){
+            if(HttpContext.Session.GetInt32("Permission") < 9){
+                return RedirectToAction("dashboard", "Dashboard");
+            }
+            if(HttpContext.Session.GetInt32("Permission") == null){
+                return RedirectToAction("Index", "Home");
+            }
             Kid kid = _context.Kids.SingleOrDefault(x => x.id == id);
             ViewBag.kid = kid;
             return View("deleteKid");
@@ -190,6 +238,12 @@ namespace AssessTracker.Controllers
         [HttpGet]
         [Route("deleteKid/{id}")]
         public IActionResult deleteKid (int id){
+            if(HttpContext.Session.GetInt32("Permission") < 9){
+                return RedirectToAction("dashboard", "Dashboard");
+            }
+            if(HttpContext.Session.GetInt32("Permission") == null){
+                return RedirectToAction("Index", "Home");
+            }
             Kid kid = _context.Kids.SingleOrDefault(x => x.id == id);
             _context.Kids.Remove(kid);
             _context.SaveChanges();
